+++ app/gui/trading/commodities_tab.py
@@ -43,8 +43,6 @@ class CommoditiesTab(BackgroundTaskMixin, QWidget):
             self.on_commodities_loaded(self.reference_service.data)
         elif self.reference_service.is_loading:
             self.on_reference_state_changed("loading")
-        else:
-            self.reference_service.ensure_loaded()
 
     def build_ui(self):
         layout = QVBoxLayout()