+++ app/gui/trading/shops_tab.py
@@ -55,8 +55,6 @@ class ShopsTab(BackgroundTaskMixin, QWidget):
             self.on_shops_loaded(self.reference_service.data)
         elif self.reference_service.is_loading:
             self.on_reference_state_changed("loading")
-        else:
-            self.reference_service.ensure_loaded()
 
     def build_ui(self):
         layout = QVBoxLayout()