+++ app/gui/trading/uex_trading_tab.py
@@ -65,8 +65,6 @@ class UEXTradingTab(BackgroundTaskMixin, QWidget):
         self.reference_service.loaded.connect(self.on_reference_loaded)
         if self.reference_service.data is not None:
             self.on_reference_loaded(self.reference_service.data)
-        else:
-            self.reference_service.ensure_loaded()
 
     def on_reference_loaded(self, data):
         update_ship_combo(self.ship_combo, data.ships)